% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BREAD-package.R
\docType{package}
\name{BREAD-package}
\alias{BREAD}
\alias{BREAD-package}
\title{BREAD: Bayesian Region-specific DNA Methylation Inference}
\description{
Targeted Bayesian inference for predefined DNA methylation regions on array
data in \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} objects. For each
user-supplied region, BREAD fits a Bayesian model, computes posterior
probabilities of directional methylation change, and classifies regions as
hypermethylated, hypomethylated, or inconclusive at user-configurable
effect-size and probability thresholds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jaimebac/BREAD}
  \item Report bugs at \url{https://github.com/jaimebac/BREAD/issues}
}

}
\author{
\strong{Maintainer}: Jaemin Park \email{jaemin.park@vai.org}

}
\keyword{internal}
