% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread.R
\name{bread}
\alias{bread}
\title{Reads a file in table format, selecting columns, subsetting rows by number and filtering them by pattern}
\usage{
bread(
  file = NULL,
  first_row = NULL,
  last_row = NULL,
  head = NULL,
  tail = NULL,
  colnames = NULL,
  colnums = NULL,
  patterns = NULL,
  filtered_columns = NULL,
  fixed = FALSE,
  meta_output = NULL,
  ...
)
}
\arguments{
\item{first_row}{Numeric. First row of the portion of the file to subset.}

\item{last_row}{Numeric. Last row of the portion of the file to subset.}

\item{head}{Numeric. How many rows starting from the first in the file.}

\item{tail}{Numeric. How many rows starting from the last in the file.}

\item{colnames}{Vector of strings. Exact names of columns to select. If both colnames and colnums are provided, colnums will be prefered.}

\item{colnums}{Vector of numeric. Columns index numbers.}

\item{patterns}{Vector of strings. One or several patterns used to filter the data from the input file. Each element of the vector should correspond to the column to be filtered. Can use regular expressions.}

\item{filtered_columns}{Vector of strings or numeric. Optional. The columns to be filtered should be indicated through their names or their index number. Each element of the vector should correspond to the pattern with which it will be filtered.}

\item{fixed}{Logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.}

\item{...}{Arguments that must be passed to data.table::fread() like "sep" or "dec".}
}
\value{
A data frame with the selected columns and the subsetted and filtered data
}
\description{
Wrapper for data.table::fread() simplifying the use of Unix commands like grep, cut and sed
on a data file \emph{before} loading it in memory. The Unix commands are automatically generated
from the arguments.
This is useful if you want to load a big file too large for your available memory
(and encounter the "cannot allocate vector of size" error) and know you can work on a
subsample. "b" stands for "big file".
This function allows to subset rows by their index number, select columns and filter
with a pattern.
}
\section{Warning}{

Best practice would probably be to load the big file in a SQL database or something.
Or not working on huge CSV files in the first place.
But if you have to, you hopefully won't have to delve into the fascinating grammar of
Unix commands.
}

\section{Arguments}{

You can mix and match the row subsetting, the filtering and the selecting of columns.
In order, the function:
\enumerate{
\item subsets the rows by their numbers (with sed). You need to input the index
number of the first and last rows you want to load in memory with fread(),
or alternatively use either the head or tail arguments to subset the first or
last rows of the file.
\item selects columns by index number or name (with cut). If both colnames and
colnums are provided, colnums will be prefered.
\item filters the data selected so far with a pattern by column (with grep). The
columns to be filtered should be indicated through their names or their index
number. Each element of the vector should correspond to the pattern with which
it will be filtered.
}
}

\examples{
bread(file = "./data/test.csv", colnums = c(1,3))
bread(file = "./data/test.csv", colnames = c("YEAR", "PRICE"), patterns = 2002, filtered_columns = "YEAR")
bread(file = "./data/test.csv", colnames = c("YEAR", "COLOR"), patterns = "red", filtered_columns = "COLOR", first_row = 10, last_row = 18)
}
\keyword{allocate}
\keyword{big}
\keyword{cut}
\keyword{file}
\keyword{select}
\keyword{size}
\keyword{vector}
