% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{classifications}
\alias{classifications}
\alias{classifications,BreadFit-method}
\title{Extract region classifications from a \link{BreadFit}}
\usage{
classifications(object, ...)

\S4method{classifications}{BreadFit}(object, ...)
}
\arguments{
\item{object}{A \link{BreadFit}.}

\item{...}{Unused.}
}
\value{
A character vector of classifications per region.
}
\description{
Extract region classifications from a \link{BreadFit}
}
