% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.beta_to_m}
\alias{.beta_to_m}
\title{Beta -> M transform, clamped away from 0/1.}
\usage{
.beta_to_m(beta, eps = 1e-06)
}
\arguments{
\item{beta}{Numeric vector of beta values.}

\item{eps}{Clamping tolerance keeping values off the 0/1 asymptotes.}
}
\value{
Numeric vector of M-values.
}
\description{
Beta -> M transform, clamped away from 0/1.
}
\keyword{internal}
