% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.m_to_beta}
\alias{.m_to_beta}
\title{M -> Beta transform.}
\usage{
.m_to_beta(m)
}
\arguments{
\item{m}{Numeric vector of M-values.}
}
\value{
Numeric vector of beta values in (0, 1).
}
\description{
M -> Beta transform.
}
\keyword{internal}
