% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_beads.R
\name{filter_low_beads}
\alias{filter_low_beads}
\title{Filters measurements below the bead threshold.}
\usage{
filter_low_beads(bg_df)
}
\arguments{
\item{bg_df}{the data frame output by rm_background function}
}
\value{
\code{filter_low_beads} returns a data frame in the same
format as rm_background, but only with measurements above the
specified bead count threshold.
}
\description{
This function filters out measurements below the user specified
bead threshold and outputs a data frame containing only measurements
with bead counts above the user specified threshold.
This is done on an individual plate basis.
}
\examples{
R code examples of how to use your function

}
