% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNinv.R
\name{FUNinv}
\alias{FUNinv}
\title{Inverse function of fit standard curve}
\usage{
FUNinv(y, par)
}
\arguments{
\item{y}{MFI value.}

\item{par}{parameters from standard curve fit.}
}
\value{
\code{FUNinv} will return estimated dilution factor
based on MFI and std curve fit.
}
\description{
Inverse function of fit standard curve
}
\examples{
R code examples of how to use your function

}
