% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concentration.R
\name{get_concentration}
\alias{get_concentration}
\title{Get estimate of Log dilution factor of individual samples.}
\usage{
get_concentration(plate_df_norm, std_curve_values, input)
}
\arguments{
\item{plate_df_norm}{a data frame that is the output of the
filter_low_beads function.}

\item{std_curve_values}{a data frame with two columns,
the first column should be named "Sample" and contain the name of
the standard curve wells as they are listed in the "Sample" column
of the original csv file. The second column  should be named
"Dilution" is the dilution factor of the standard curve sample
as a numeric fraction e.g. 1/1000.}

\item{input}{a string either "MFI" or "bgMFI" to designate whether
the standard curve should be fit with raw MFI values or background
corrected MFI values}
}
\value{
\code{get_concentration} returns a data frame containing
all variables from the filter_low_beads function plus a new variable
with the estimated log concentration obtained by fitting the standard
curve.
}
\description{
This function fits standard curves by estimating the
parameters of a five point logistic curve then using these
estimated parameters, takes the inverse of this logistic curve
to estimate the log concentration of samples. The user can specify
whether the input should be raw or background corrected MFI values.
}
\examples{
R code examples of how to use your function


}
