% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bread.R
\name{make_bread}
\alias{make_bread}
\title{Make bread from grains, yeast, water and salt}
\usage{
make_bread(grains, yeast, water, salt)
}
\arguments{
\item{grains}{A scalar representing the amount of grains.}

\item{yeast}{A scalar representing the amount of yeast.}

\item{water}{A scalar representing the amount of water.}

\item{salt}{A scalar representing the amount of salt.}
}
\value{
A scalar representing the amount of bread.
}
\description{
Make bread from grains, yeast, water and salt
}
\examples{
make_bread(grains = 1, yeast = 1, water = 1.5, salt = 3)

}
