% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_focaccia.R
\name{make_focaccia}
\alias{make_focaccia}
\title{Make focaccia from grains, yeast, water and salt}
\usage{
make_focaccia(grains, yeast, water, salt)
}
\arguments{
\item{grains}{A scalar representing the amount of grains.}

\item{yeast}{A scalar representing the amount of yeast.}

\item{water}{A scalar representing the amount of water.}

\item{salt}{A scalar representing the amount of salt.}
}
\value{
A scalar representing the amount of focaccia.
}
\description{
Make focaccia from grains, yeast, water and salt
}
\examples{
make_focaccia(grains = 1, yeast = 1, water = 1.5, salt = 3)

}
