% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_between_plates.R
\name{normalize_between_plates}
\alias{normalize_between_plates}
\title{Performs simple between-plate normalization.}
\usage{
normalize_between_plates(all_plates_df)
}
\arguments{
\item{all_plates_df}{the result of binding together (rbind)
all plate oututs from the get_concentration function into one
data frame.}
}
\value{
\code{normalize_between_plates} returns a data frame containing
all the same outputs as the get_concentration function as well
as a new output of normalized log concentration, information
from all plates is combined into one large data frame.
}
\description{
Using measurements that are repeated between plates (control samples).
This function performs a linear regression with log estimate concentration,
the output of the get_concentration function, as the outcome and
plate, antigen, and sample name as regressors. The estimated plate effects
are then subtracted from the log concentrations of all samples
on respective plates to produce a normalized log concentration.
}
\examples{
R code examples of how to use your function

}
