% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beadcount.R
\name{plot_beadcount}
\alias{plot_beadcount}
\title{Shows a histogram of beadcounts.}
\usage{
plot_beadcount(bg_df, bead_threshold)
}
\arguments{
\item{bg_df}{the data frame output by rm_background function}

\item{bead_threshold}{a numeric value indicating the minimum number of
beads for a measurement to be considered in downstream analysis.}
}
\value{
\code{plot_bead_count} returns histograms of beadcounts for a
particular plate.
}
\description{
This function takes in the output of rm_background and plots the
beadcounts in the form of a histogram across all samples and antigens
for a particular plate. It also shows the cutoff for low bead counts
as specified by the user. This is done on an individual plate basis.
}
\examples{
R code examples of how to use your function

}
