% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_region_data}
\alias{plot_region_data}
\title{Raw region-level values by contrast group}
\usage{
plot_region_data(fit, region_id)
}
\arguments{
\item{fit}{A \link{BreadFit}.}

\item{region_id}{Single region ID.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Boxplot + jitter of the summarized region values for a single region,
grouped by the first variable in the design formula.
}
\examples{
suppressPackageStartupMessages(library(SummarizedExperiment))

se  <- readRDS(system.file("extdata", "vitc_ag06561.rds", package = "BREAD"))
reg <- readRDS(system.file("extdata", "vitc_regions.rds", package = "BREAD"))
se_ctrl <- se[, se$condition == "ctrl"]

fit <- fit_bread(se_ctrl, reg, ~ passage)

res <- results(fit)
rid <- res$region_id[which(res$classification == "hypermethylated")[1]]
if (is.na(rid)) rid <- res$region_id[1]

# `region_id` must be a single region.
plot_region_data(fit, rid)
}
