% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_std_curves.R
\name{plot_std_curves}
\alias{plot_std_curves}
\title{Plots standard curves.}
\usage{
plot_std_curves(plate_bg_df, std_curve_values, input)
}
\arguments{
\item{plate_bg_df}{a dataframe that is the output of filter_low_beads.}

\item{input}{a string either "bgMFI" or "MFI" indicating whether
to visualize background corrected MFI or raw MFI on the y-axis.}

\item{standard_curve_values}{a data frame with two columns,
the first column should be named "Sample" and contain the name of
the standard curve wells as they are listed in the "Sample" column
of the original csv file. The second column  should be named
"Dilution" is the dilution factor of the standard curve sample
as a numeric fraction e.g. 1/1000.}
}
\value{
\code{plot_std_curves} returns plots of standard curves.
}
\description{
This function plots standard curves for individual plates on the
log-log scale (log dilution factor on the x-axis, and log
MFI on the y-axis).Users can specify whether they would like to
visualize the background corrected MFI or the raw MFI.
}
\examples{
R code examples of how to use your function

}
