% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_beadcounts.R
\name{table_beadcounts}
\alias{table_beadcounts}
\title{Returns a data frame summarizing low bead counts.}
\usage{
table_beadcounts(bg_df)
}
\arguments{
\item{bg_df}{the data frame output by rm_background function}
}
\value{
\code{table_beadcounts} returns a table of low beadcounts
for a particular plate.
}
\description{
This function takes in the output of rm_background and returns
a data frame summarizing how low bead counts are distributed
across sample types and antigens. This is done on an individual plate
basis.
}
\examples{
R code examples of how to use your function

}
