% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BREAD-package.R
\docType{package}
\name{BREAD-package}
\alias{BREAD}
\alias{BREAD-package}
\title{BREAD: This package allows users to read in raw data from multiplex bead experiments on the Luminex platform and compare the impacts of various pre-processing steps on their data.}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sberube3/BREAD}
  \item Report bugs at \url{https://github.com/sberube3/BREAD/issues}
}

}
\author{
\strong{Maintainer}: Sophie Berube \email{sberube1@ufl.edu}

}
\keyword{internal}
