% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{classifications}
\alias{classifications}
\alias{classifications,BreadFit-method}
\title{Extract region classifications from a \link{BreadFit}}
\usage{
classifications(object, ...)

\S4method{classifications}{BreadFit}(object, ...)
}
\arguments{
\item{object}{A \link{BreadFit}.}

\item{...}{Unused.}
}
\value{
A named character vector of classifications per region
(names are region IDs, values are one of \code{"hypermethylated"},
\code{"hypomethylated"}, \code{"inconclusive"}).
}
\description{
Extract region classifications from a \link{BreadFit}
}
\examples{
suppressPackageStartupMessages(library(SummarizedExperiment))

se  <- readRDS(system.file("extdata", "vitc_ag06561.rds", package = "BREAD"))
reg <- readRDS(system.file("extdata", "vitc_regions.rds", package = "BREAD"))
se_ctrl <- se[, se$condition == "ctrl"]

fit <- fit_bread(se_ctrl, reg, ~ passage)

cls <- classifications(fit)
head(cls)
table(cls)
}
