% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bread.R
\name{make_bread}
\alias{make_bread}
\title{Make bread}
\usage{
make_bread(grains, yeast, water, salt)
}
\arguments{
\item{grains}{Amount of grains (numeric).}

\item{yeast}{Amount of yeast (numeric).}

\item{water}{Amount of water (numeric).}

\item{salt}{Amount of salt (numeric).}
}
\value{
Amount of bread.
}
\description{
Make bread
}
\examples{
make_bread(
  grains = 1,
  yeast = 2,
  water = 3,
  salt = 4
)
}
