% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_focaccia.R
\name{make_focaccia}
\alias{make_focaccia}
\title{Make focaccia}
\usage{
make_focaccia(grains, yeast, water, salt)
}
\arguments{
\item{grains}{Amount of grains (numeric).}

\item{yeast}{Amount of yeast (numeric).}

\item{water}{Amount of water (numeric).}

\item{salt}{Aùount of salt (numeric).}
}
\value{
Amount of focaccia (numeric).
}
\description{
Make focaccia
}
\examples{
make_focaccia(
  grains = 1,
  yeast = 2,
  water = 3,
  salt = 4
)
}
