% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{posterior_draws}
\alias{posterior_draws}
\alias{posterior_draws,BreadFit-method}
\title{Extract posterior draws for the region-level effect}
\usage{
posterior_draws(object, region_id = NULL, ...)

\S4method{posterior_draws}{BreadFit}(object, region_id = NULL, ...)
}
\arguments{
\item{object}{A \link{BreadFit}.}

\item{region_id}{Optional region ID to subset.}

\item{...}{Unused.}
}
\value{
A tidy data frame of draws.
}
\description{
Extract posterior draws for the region-level effect
}
